[LOCAL][BUGFIX] Make compliant with gcc warning and pedantic flags

+++ 0x04-more_functions_nested_loops/2-mul.c
@@ -10,7 +10,7 @@
  */
 int mul(int a, int b)
 {
-	long long int a1, b1;
+	long int a1, b1;
 
 	a1 = a;
 	b1 = b;
